+++ frontend/eda/UserAccess/Users/hooks/useDeleteUser.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { compareStrings, useBulkConfirmation } from '../../../../../framework';
 import { requestDelete } from '../../../../Data';
-import { idKeyFn } from '../../../../hub/useHubView';
+import { idKeyFn } from '../../../../galaxy/useHubView';
 import { EdaUser } from '../../../interfaces/EdaUser';
 import { useUserColumns } from './useUserColumns';
 import { API_PREFIX } from '../../../constants';
