+++ frontend/eda/execution-environments/hooks/useDeleteExecutionEnvironments.tsx
@@ -2,7 +2,7 @@ import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { compareStrings, useBulkConfirmation } from '../../../../framework';
 import { requestDelete } from '../../../Data';
-import { idKeyFn } from '../../../hub/useHubView';
+import { idKeyFn } from '../../../galaxy/useHubView';
 import { EdaExecutionEnvironment } from '../../interfaces/EdaExecutionEnvironment';
 import { useExecutionEnvironmentColumns } from './useExecutionEnvironmentColumns';
 import { API_PREFIX } from '../../constants';
