+++ frontend/galaxy/administration/remote-registries/RemoteRegistries.tsx
