+++ frontend/galaxy/administration/remote-registries/RemoteRegistry.tsx
