+++ frontend/galaxy/administration/remote-registries/hooks/useRemoteRegistriesActions.tsx
