+++ frontend/galaxy/administration/remote-registries/hooks/useRemoteRegistriesColumns.tsx
