+++ frontend/galaxy/administration/remote-registries/hooks/useRemoteRegistryActions.tsx
