+++ frontend/galaxy/automation-content/collections/Collection.tsx
