+++ frontend/galaxy/automation-content/collections/CollectionDetails.tsx
