+++ frontend/galaxy/automation-content/collections/Collections.tsx
