+++ frontend/galaxy/automation-content/collections/FeaturedCollections.tsx
