+++ frontend/galaxy/automation-content/collections/UploadCollection.tsx
