+++ frontend/galaxy/automation-content/collections/hooks/useCollectionActions.tsx
