+++ frontend/galaxy/automation-content/collections/hooks/useCollectionColumns.tsx
