+++ frontend/galaxy/automation-content/collections/hooks/useCollectionFilters.tsx
