+++ frontend/galaxy/automation-content/collections/hooks/useCollections.tsx
