+++ frontend/galaxy/automation-content/collections/hooks/useCollectionsActions.tsx
