+++ frontend/galaxy/automation-content/collections/hooks/useDeleteCollections.tsx
