+++ frontend/galaxy/automation-content/collections/hooks/useDeprecateCollections.tsx
