+++ frontend/galaxy/automation-content/namespaces/Namespace.tsx
