+++ frontend/galaxy/automation-content/namespaces/NamespaceDetails.tsx
