+++ frontend/galaxy/automation-content/namespaces/Namespaces.tsx
