+++ frontend/galaxy/automation-content/namespaces/hooks/useNamespaces.tsx
