+++ frontend/galaxy/automation-content/namespaces/hooks/useNamespacesColumns.tsx
