+++ frontend/galaxy/useHubView.tsx
